+++ server/lib/controllers/commentController.js
@@ -19,8 +19,8 @@ commentController.getAll = async (req, res) => {
 commentController.post = async (req, res) => {
   const comment = new _Comment.default({
     articleID: req.body.articleID,
-    name: req.body.name,
-    email: req.body.email,
+    userID: req.user._id,
+    name: req.user.name,
     commentBody: req.body.commentBody
   });
   await comment.save();
@@ -42,6 +42,12 @@ commentController.getOne = async (req, res) => {
 };
 
 commentController.delete = async (req, res) => {
+  if (!(req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com")) {
+    return res.status(401).send({
+      error: 'Unauthorized action.'
+    });
+  }
+
   try {
     await _Comment.default.deleteOne({
       _id: req.params.id
