+++ server/src/controllers/likeController.js
@@ -3,7 +3,9 @@ import Like from "../models/Like";
 const likeController = {};
 
 likeController.getAll = async (req, res) => {
-
+  if(!(req.user._id == req.params.id || (req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com" ))){
+      return res.status(401).send({error:'Unauthorized action.'});
+  }
   const likes = await Like.find();
   res.send(likes);
 
@@ -11,25 +13,33 @@ likeController.getAll = async (req, res) => {
 
 likeController.post = async (req, res) => {
 
-  const storedLike = await Like.findOne({ email: req.body.email });
+  const storedLike = await Like.findOne({ 
+    articleID: req.body.articleID,
+    userID : req.user._id  
+  });
   if(storedLike == null){
     const like = new Like({
       articleID: req.body.articleID,
-      email : req.body.email,
-      likeBool: req.body.likeBool
+      userID : req.user._id
     });
   
     await like.save();
     res.send(like);
   }
   else{
-    await Like.deleteOne({ email: req.body.email });
+    await Like.deleteOne({ 
+      articleID: req.body.articleID,
+      userID : req.user._id 
+    });
     res.status(204).send();
   }
   
 };
 
 likeController.getOne = async (req, res) => {
+  if(!(req.user._id == req.params.id || (req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com" ))){
+    return res.status(401).send({error:'Unauthorized action.'});
+  }
   try {
 
     const like = await Like.find({ articleID: req.params.id });
@@ -45,6 +55,9 @@ likeController.getOne = async (req, res) => {
 
 
 likeController.delete = async (req, res) => {
+  if(!(req.user._id == req.params.id || (req.user.membership == "admin" || req.user.email == "smbonimpa2011@gmail.com" ))){
+    return res.status(401).send({error:'Unauthorized action.'});
+  }
   try {
 
     await Like.deleteOne({ _id: req.params.id });
